grabbed Main and combined Util
